sqs_driver/requirements.txt
boto3==1.34.63
botocore==1.34.63
jmespath==1.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
urllib3==1.26.18
